# Conflicts:
#	src/main/java/org/choongang/thesis/services/ThesisInfoService.java
#	src/main/java/org/choongang/thesis/services/WishListService.java

+++ src/main/java/org/choongang/thesis/services/WishListService.java
@@ -2,6 +2,8 @@ package org.choongang.thesis.services;
 
 import com.querydsl.core.BooleanBuilder;
 import lombok.RequiredArgsConstructor;
+import org.choongang.global.Utils;
+import org.choongang.global.exceptions.BadRequestException;
 import org.choongang.member.MemberUtil;
 import org.choongang.thesis.entities.QWishList;
 import org.choongang.thesis.entities.WishList;
@@ -20,25 +22,25 @@ public class WishListService {
 
     private final MemberUtil memberUtil;
     private final WishListRepository wishListRepository;
+    private final Utils utils;
 
     public void add(Long tid){
-//        if(!memberUtil.isLogin()){
-//            return;
-//        }
+        if(!memberUtil.isLogin()){
+            throw new BadRequestException(utils.getMessage("Login.Required"));
+        }
         WishList wishList = WishList.builder()
                 .tid(tid)
-//                .email(memberUtil.getMember().getEmail())
-                .email("testuser1@email.com")
+                .email(memberUtil.getMember().getEmail())
                 .build();
         System.out.println("wishList : "+wishList);
         wishListRepository.saveAndFlush(wishList);
     }
 
     public void remove(Long tid){
-//        if(!memberUtil.isLogin()){
-//            return;
-//        }
-        WishListId wishListId = new WishListId(tid, "testuser1@email.com");
+        if(!memberUtil.isLogin()){
+            throw new BadRequestException(utils.getMessage("Login.Required"));
+        }
+        WishListId wishListId = new WishListId(tid,memberUtil.getMember().getEmail());
         wishListRepository.deleteById(wishListId);
         wishListRepository.flush();
         System.out.println("위시리스트 논문 삭제");
